requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
email-validator==2.2.0
httpx==0.27.0
redis[hiredis]==5.0.7
websockets==12.0
celery==5.4.0
aiocache==0.12.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-multipart==0.0.9
ruff==0.6.5
pytest==8.3.2

